FIX 주문 및 결제 로직 중, 쿠폰 관련 에러 헨들링 수정
